altikelime/requirements.txt
appdirs==1.4.3
appnope==0.1.0
astroid==2.1.0
attrs==18.2.0
backcall==0.1.0
bandit==1.5.1
black==18.9b0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.2
decorator==4.3.2
django>=2.1.6
django-debug-toolbar==1.11
django-extensions==2.1.4
django-filter==2.1.0
djangorestframework==3.9.1
flake8==3.6.0
flake8-bandit==2.0.0
flake8-blind-except==0.1.1
flake8-bugbear==18.8.0
flake8-builtins==1.4.1
flake8-polyfill==1.0.2
flake8-print==3.1.0
flake8-quotes==1.0.0
flake8-string-format==0.2.3
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
ipython==7.1.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.2
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
parso==0.3.4
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.3.1
pylint==2.2.2
pytz==2018.9
pyyaml>=4.2b1
requests==2.21.0
six==1.12.0
smmap2==2.0.5
sqlparse==0.2.4
stevedore==1.30.0
toml==0.10.0
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.1
